Rename 'individualApplication' to 'applicationView'
