+++ src/app/App.tsx
@@ -12,7 +12,7 @@ import LoginPage from '../features/auth/loginPage/LoginPage';
 import Page from './page/Page';
 import PrivateRoute from '../features/auth/privateRoute/PrivateRoute';
 import ApplicationsPage from '../features/applicationList/ApplicationsPageContainer';
-import IndividualApplicationPage from '../features/individualApplication/IndividualApplicationPageContainer';
+import IndividualApplicationPage from '../features/applicationView/IndividualApplicationPageContainer';
 import PricingPage from '../features/pricing/PricingPageContainer';
 import CallbackPage from '../features/auth/callbackPage/CallbackPage';
