+++ src/features/applicationView/IndividualApplicationPage.tsx
