+++ src/features/applicationView/individualApplicationPage.module.scss
