+++ src/features/applicationView/offerCard/OfferCard.tsx
