+++ src/features/applicationView/offerCard/offerCard.module.scss
