+++ src/features/applicationView/utils.ts
